test2/requirements.txt
aiohttp>=3.9.5
aiomqtt>=2.1.0
colorlog>=6.8.2
getmac>=0.9.4
paho-mqtt>=2.1.0
PyYAML>=6.0.1
rich>=13.7.1
SQLAlchemy>=2.0.30
starlette>=0.37.2
uvicorn>=0.29.0
yarl>=1.9.4
zeroconf>=0.132.2
requests>=2.32.3
PyYAML>=6.0.1